Fix cross-platform compilation issues for modern toolchains
- Fix SIMD macro definitions to work on non-x86 platforms
- Add GitHub Actions CI/CD with vcpkg for multi-platform builds
- Fix Windows build issues (process.h includes, thread functions)
- Fix macOS x64 SIMD prefetch constraints
- Fix LibGD conditional compilation for platforms without LibGD
- Add HAVE_LIBGD guards around GD-dependent code in GfxSaveSurface and MapLoader_LieroX
- Fix process.h include ordering in HawkNL thread.c for Windows

+++ src/common/MapLoader_Teeworlds.cpp
@@ -1104,7 +1104,7 @@ Result TWGroup::read(ML_Teeworlds* l, char* p, char* end) {
 	else
 		name = "";
 
-	if(p > end) return "group item data is invalid, read behind end by " + itoa(p - end) + " bytes";
+	if(p > end) return "group item data is invalid, read behind end by " + itoa((int)(p - end)) + " bytes";
 	return true;
 }
